// src/app/shared/components/utils/input/input.component.interface.ts
export enum InputType {
	BUTTON = 'button',
	CHECKBOX = 'checkbox',
	COLOR = 'color',
	DATE = 'date',
	DATETIME_LOCAL = 'datetime-local',
	EMAIL = 'email',
	FILE = 'file',
	HIDDEN = 'hidden',
	IMAGE = 'image',
	MONTH = 'month',
	NUMBER = 'number',
	PASSWORD = 'password',
	RADIO = 'radio',
	RANGE = 'range',
	RESET = 'reset',
	SEARCH = 'search',
	SUBMIT = 'submit',
	TEL = 'tel',
	TEXT = 'text',
	TIME = 'time',
	URL = 'url',
	WEEK = 'week',
}

export enum InputTheme {
	ROUNDED_DEFAULT = 'ROUNDED_DEFAULT',
}